Small refactoring to EXECUTE-COMMAND.
The substantial change is to join the items with space separator so that
at the invocations we don't need to put spaces around the commands. This
is both for aesthetics and for the future where the commands may not be
specifying words on a command line; this change hides that
implementation detail.

The smaller change is to rename the arguments of the function. Having a
function called EXECUTE-COMMAND which didn't take any parameters called
COMMAND seemed odd.